added folder: the new timer and pc, player score
in pyonpyon.v, included the timer and pc score counter (rate divider and speed selection).
in count_down_player.v , that is the code for player score. But not sure how it would be implemented in the main code. This is the code when the count_down_player receive a signal called "correct" and it will check every positive clock.